% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{correct_leaf_area}
\alias{correct_leaf_area}
\title{calculate leaf area sandwitched by the chamber}
\usage{
correct_leaf_area(data, width_mm, model = "6400-40")
}
\arguments{
\item{data}{gas-exchange data}

\item{width_mm}{leaf width in mm}

\item{model}{chamber used for the measurement}
}
\description{
calculate leaf area sandwitched by the chamber
}
