% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/li6400xt-package.R
\docType{package}
\name{li6400xt-package}
\alias{li6400xt}
\alias{li6400xt-package}
\title{li6400xt: Preprocess Gas Exchange Data for R workflow}
\description{
Read gas exchange data obtained by LI-6400XT.
}
\author{
\strong{Maintainer}: Keach Murakami \email{keach.murakami@gmail.com} (\href{https://orcid.org/0000-0001-8150-9535}{ORCID})

}
\keyword{internal}
