% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_li6400xt.R
\name{read_licor}
\alias{read_licor}
\title{read gas-exchange data from LI6400XT}
\usage{
read_licor(file, inherit_pam = FALSE)
}
\arguments{
\item{file}{path}

\item{inherit_pam}{logical; if \code{TRUE} xxx}
}
\description{
read gas-exchange data from LI6400XT
}
