% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{recalculate}
\alias{recalculate}
\title{recalculate gas exchange parameters by correcting leaf area}
\usage{
recalculate(data, corrected_area, default_area)
}
\arguments{
\item{data}{gas-exchange data}

\item{corrected_area}{corrected leaf area in cm2}

\item{default_area}{default leaf area in cm2}
}
\description{
recalculate gas exchange parameters by correcting leaf area
}
