% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_li6400xt.R
\name{track_changes}
\alias{track_changes}
\title{extract control procedure from change logs}
\usage{
track_changes(data, track = "LCF", variable = "PPFD")
}
\arguments{
\item{data}{output of \code{read_licor}}

\item{track}{tracked operation}

\item{variable}{name of new variable}
}
\description{
extract control procedure from change logs
}
