% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_li6400xt.R
\name{track_changes}
\alias{track_changes}
\title{extract control procedure from change logs}
\usage{
track_changes(data, track_variable = "ParIn", name_variable = "PPFD")
}
\arguments{
\item{data}{output of \code{read_licor}}

\item{track_variable}{variable tracked}

\item{name_variable}{name of variable in the output tibble}
}
\description{
extract control procedure from change logs
}
